Bunch of warning fixups.
imported from: dustin@spy.net--projects-2005/photo--main--4.0--patch-141

+++ src/java/net/spy/photo/Votes.java
@@ -28,8 +28,8 @@ public class Votes extends SpyObject {
 	 */
 	public void add(Vote v) {
 		store.put(v.getUser().getId(), v);
-		total += (double)v.getVote();
-		average=total / (double)store.size();
+		total += v.getVote();
+		average=total / store.size();
 	}
 
 	/**
